feat: add `youtube` parameter to `buildPageTargeting` (#708)
* Add `youtube` parameter to `buildPageTargeting`

When `youtube` is set to `false`, `buildPageTargeting` pulls the regular
permutive segments (i.e. those stored in localstorage under the key
`_papns`). This was the default behaviour before this change.

When `youtube` is set to `true`, `buildPageTargeting` pulls the
youtube-specific permutive segments (i.e. those stored in localstorage
under the key `_pdfps`)

* Use `buildPageTargeting` in `buildAdsConfig`

* Remove stray `.only`

Oops, this was my fault, I left it in by mistake. It was limiting the
value of this spec!