Remove the non refreshable line item fallback (#921)
* Remove usage of fetchNonRefreshableLineItems

Remove any usage of `fetchNonRefreshableLineItems` switch.

We previously had the switch in order to fallback to the old
method of retrieving the non refreshable line item ids. The
new method has been stable for a long time, so it's safe to
remove the switch and the fallback now.

* Ensure the final should-refresh test fails now

Once we remove the fallback, the final test should fail. However,
it passes.

This is because we're not properly resetting the window in between
tests. This should fix it.

* Update should-refresh test and test newer method

* Add a changeset