Revert "Only run Prebid in GDPR regions if all TCF Prebid vendors have consent" (#1828)
We don't actually want to do this. If there are new vendors added to the vendor list and a user is not prompted to re-consent, this will prevent us being able to run Prebid at all for those users. Since users are not prompted to re-consent for every new vendor added, this poses a revenue risk. It will be better to manage this on a more granular basis