Report failure at earlier stage in Playwright in E2E workflow (#2332)
Sometimes there is no `main` tag available on DCR container images in the Github container registry. This happens occasionally and the fix is to run CI on DCR main branch to get this up to date. The previous PR for this https://github.com/guardian/commercial/issues/1736 did not manage to raise failures early as the `outputs` weren't defined due to the previous step exiting due to errors. This PR fixes this issue and uses `outcomes` instead to detect failure in the previous step.